Move email validation to a reusable concern
